warning and debug output cleanup

+++ lib-polymer/procedural_mesh.hpp
@@ -573,7 +573,7 @@ namespace polymer
         
         auto curve = bezier_spline(float3(0.0f, 0.0f, 0.0f), float3(0.667f, 0.25f, 0.0f), float3(1.33f, 0.25f, 0.0f), float3(2.0f, 0.0f, 0.0f));
         
-        const int numSegments = curve.num_steps();
+        const int numSegments = static_cast<int>(curve.num_steps);
         const int numSlices = numSegments + 1;
         const int numVerts = 2 * numSlices;
